refactor: split provider page seams
