+++ frontend/src/App.tsx
@@ -6,8 +6,10 @@ import { useToasts } from "./components/PageToast";
 import { changeAppLocale } from "./i18n";
 import { cx } from "./lib/format";
 import { historyIssueKey } from "./lib/history";
+import { addDays, endOfDay, startOfDay } from "./lib/dateRange";
 import { api, messageFromError } from "./lib/tauri";
 import contract from "./lib/ui-contract.json";
+import { isUpdateInstallActive, updateInstallToastText } from "./lib/updateStatus";
 import type {
   AppFlavorInfo,
   AppStatus,
@@ -86,18 +88,6 @@ function defaultUsageWindow(): UsageQueryWindow {
     startTs: addDays(end, -6).toISOString(),
     endTs: endOfDay(end).toISOString(),
   };
-}
-
-function startOfDay(date: Date) {
-  return new Date(date.getFullYear(), date.getMonth(), date.getDate());
-}
-
-function endOfDay(date: Date) {
-  return new Date(date.getFullYear(), date.getMonth(), date.getDate(), 23, 59, 59, 999);
-}
-
-function addDays(date: Date, days: number) {
-  return new Date(date.getFullYear(), date.getMonth(), date.getDate() + days);
 }
 
 function runtimeCache<T>(data: T | null = null): RuntimeCache<T> {
@@ -1129,45 +1119,6 @@ export default function App() {
       />
     </div>
   );
-}
-
-function isUpdateInstallActive(status: AppUpdateInstallStatus | null | undefined) {
-  return Boolean(
-    status &&
-      (status.phase === "checking" ||
-        status.phase === "downloading" ||
-        status.phase === "installing" ||
-        status.phase === "restarting"),
-  );
-}
-
-function updateInstallProgressPercent(status: AppUpdateInstallStatus) {
-  if (status.phase !== "downloading" || !status.total_bytes || status.total_bytes <= 0) {
-    return null;
-  }
-  return Math.max(0, Math.min(100, Math.round((status.downloaded_bytes / status.total_bytes) * 100)));
-}
-
-function updateInstallToastText(
-  status: AppUpdateInstallStatus,
-  t: (key: string, options?: Record<string, unknown>) => string,
-) {
-  if (status.phase === "checking") {
-    return t("settings.checkingUpdates");
-  }
-  if (status.phase === "downloading") {
-    const percent = updateInstallProgressPercent(status);
-    return percent === null
-      ? t("settings.downloadingUpdate")
-      : t("settings.downloadingUpdateProgress", { percent });
-  }
-  if (status.phase === "installing" || status.phase === "restarting") {
-    return t("settings.installingUpdateRestarting");
-  }
-  if (status.phase === "failed") {
-    return t("settings.updateInstallFailed", { message: status.message });
-  }
-  return status.target_version ? t("settings.installingUpdateRestarting") : t("settings.updateInstallUnavailable");
 }
 
 function failedUpdateInstallStatus(
